+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 - add `edge`
 - simplified and generic edge iterator (now a Generator)
 - add `edges(g, v)` and `edges(g, vertices)`
+- induced_subgraph -> subgraph